auth-server: verify nullifier spent corresponds with bundle being waited
